requirements.txt
aiohappyeyeballs==2.4.6aiohttp==3.11.12aiosignal==1.3.2annotated-types==0.7.0anthropic==0.46.0anyio==4.8.0asttokens==2.4.1attrs==25.1.0babel==2.17.0backoff==2.2.1beautifulsoup4==4.13.3blinker==1.9.0boto3==1.36.25botocore==1.36.25browser-use==0.1.37cachetools==5.5.2certifi==2025.1.31charset-normalizer==3.4.1click==8.2.1colorama==0.4.6comm==0.2.2contourpy==1.3.2courlan==1.3.2cycler==0.12.1dateparser==1.2.1debugpy==1.8.7decorator==5.1.1defusedxml==0.7.1Deprecated==1.2.18distro==1.9.0executing==2.1.0filetype==1.2.0fireworks-ai==0.15.12Flask==3.1.1fonttools==4.57.0frozenlist==1.5.0gdown==5.2.0google-ai-generativelanguage==0.6.15google-api-core==2.24.1google-api-python-client==2.161.0google-auth==2.38.0google-auth-httplib2==0.2.0google-generativeai==0.8.4googleapis-common-protos==1.68.0greenlet==3.1.1grpcio==1.67.1grpcio-status==1.67.1gunicorn==21.2.0h11==0.14.0html2text==2024.2.26htmldate==1.9.3httpcore==1.0.7httplib2==0.22.0httpx==0.28.1httpx-sse==0.4.0httpx-ws==0.7.1idna==3.10importlib_metadata==8.5.0ipykernel==6.29.5ipython==8.28.0itsdangerous==2.2.0jedi==0.19.1Jinja2==3.1.6jiter==0.8.2jmespath==1.0.1joblib==1.4.2jsonpatch==1.33jsonpointer==3.0.0jupyter_client==8.6.3jupyter_core==5.7.2jusText==3.0.1kiwisolver==1.4.8langchain==0.3.14langchain-anthropic==0.3.3langchain-aws==0.2.13langchain-core==0.3.37langchain-fireworks==0.2.7langchain-google-genai==2.0.8langchain-ollama==0.2.2langchain-openai==0.3.1langchain-text-splitters==0.3.6langsmith==0.2.11lmnr==0.4.61lxml==5.3.1lxml-html-clean==0.4.1MainContentExtractor==0.0.4markdownify==0.14.1MarkupSafe==3.0.2matplotlib==3.10.1matplotlib-inline==0.1.7monotonic==1.6multidict==6.1.0nest-asyncio==1.6.0numpy==1.26.4ollama==0.4.7openai==1.63.2opentelemetry-api==1.30.0opentelemetry-exporter-otlp-proto-common==1.30.0opentelemetry-exporter-otlp-proto-grpc==1.30.0opentelemetry-exporter-otlp-proto-http==1.30.0opentelemetry-instrumentation==0.51b0opentelemetry-instrumentation-langchain==0.38.7opentelemetry-instrumentation-requests==0.51b0opentelemetry-instrumentation-sqlalchemy==0.51b0opentelemetry-instrumentation-threading==0.51b0opentelemetry-instrumentation-urllib3==0.51b0opentelemetry-proto==1.30.0opentelemetry-sdk==1.30.0opentelemetry-semantic-conventions==0.51b0opentelemetry-semantic-conventions-ai==0.4.2opentelemetry-util-http==0.51b0orjson==3.10.15packaging==24.1pandas==2.2.3parso==0.8.4pillow==11.1.0platformdirs==4.3.6playwright==1.50.0posthog==3.14.2prompt_toolkit==3.0.48propcache==0.3.0proto-plus==1.26.0protobuf==5.29.3psutil==6.1.0pure_eval==0.2.3pyasn1==0.6.1pyasn1_modules==0.4.1pydantic==2.10.6pydantic-core==2.27.2pyee==12.1.1Pygments==2.18.0pyparsing==3.2.1python-dateutil==2.9.0.post0python-dotenv==1.0.1pytz==2024.2PyYAML==6.0.2pyzmq==26.2.0regex==2024.11.6requests==2.32.3requests-toolbelt==1.0.0rsa==4.9s3transfer==0.11.2scikit-learn==1.6.1scipy==1.15.2seaborn==0.13.2six==1.16.0sniffio==1.3.1soupsieve==2.6SQLAlchemy==2.0.38stack-data==0.6.3tenacity==9.0.0threadpoolctl==3.6.0tiktoken==0.9.0tld==0.13tornado==6.4.1tqdm==4.67.1trafilatura==2.0.0traitlets==5.14.3typing_extensions==4.12.2tzdata==2024.2tzlocal==5.3uritemplate==4.1.1urllib3==2.3.0wcwidth==0.2.13Werkzeug==3.1.3wrapt==1.17.2wsproto==1.2.0xgboost==3.0.2yarl==1.18.3zipp==3.21.0